PluginKit: API Break - unify update/wait callbacks and their methods.
This reduces the difference between coding a sync and an async device,
and allows both a simpler plugin API and simpler internals.

Plugins written using PluginKit will need to be modified, replacing
calls to osvrDeviceSyncRegisterUpdateCallback and
osvrDeviceAsyncStartWaitLoop with osvrDeviceRegisterUpdateCallback

Sorry for any inconvenience - easier to do this now than later.