Fix a potentially bad bug in the treenode: use indices to iterate through children.
In the non-const visitor, it's conceivable that a visitor might
add or remove a child, and apparently at least MSVC uses iterators
to implement range-based for loops. Running osvr_print_tree in a debug
build triggered a "safe iterator" assert which I finally figured out
was because of being able to modify the container we were iterating over.