Port Analog to using DeviceInterfaceBase.
This is part of the initial work for #53 - this same transformation
needs to be done to the remaining device interfaces: this will mean
they'll still "take" a dev token, but won't actually use it.

Then, once we are clear for an API break, we can remove those unused
arguments from the signatures.