Fix strange build error related to OpenCV.
If we just add an interface include directory for a build interface,
on the imaging targets, without an install one, we get an error. However,
if we adjust the imported targets to have an include directory, and then
only link against those in build interface, then it works fine.

Not sure if this is a CMake bug or a bug in our usage.