Bug fix for NetworkImagingRemoteHandler.
Callbacks were triggered before ownership of the image buffer was transferred to the context. This made it impossible to free the image buffer immediately in the callback - instead you'd have to wait until after the callback is triggered to free the buffer. Now you can free the buffer immediately after use inside the callback.