VideoIMUFusion basic initial working version.
IMU updates trigger a full pose report, we are just using the orientation
from the IMU directly. Position from the video tracker, moved to room
space by the initial several frames of data (filtered), is put through a
1-euro filter.

Think I've got some math in the wrong order here, rotation is giving me
"sweeping" translation that we don't want, but this is a place to start.