Refactor how RemoteHandlers interact with interfaces. (Internal API break!)
In-progress additional interface classes will need to be updated:
triggerCallbacks no longer automatically sets the state.

This commit includes changes to use an additional class instead of a
reference to a list of client interfaces directly, so existing interfaces
classes should all work, and their implementation is now simpler.

Also, as part of this change, I believe that we should have ref-counting
right for images: fixes #216