Fix Video-based tracker with OpenCV 3.x.
Turns out that OpenCV 3.0 changed the semantic signature of solvePnPRansac,
while having it type-compatible: the minimum number of inliers was replaced
by a 0 to 1 "confidence" parameter. Lacking any information from their docs
on what the change meant or how to reasonably convert code from one convention
to the other, the OpenCV 3 path now just uses the default confidence of 0.99,
labeled with a @todo, which should be fine especially since it's just used
to bootstrap the SCAAT filter now.