Don't give typesafeid a reference accessor by default, keeping them safer.
This makes them more type-safe by default (you can't just assign
to .value() without explicitly permitting it for a given type) but
does break some usages of them that relies on this. The ones
already in master (in StringIds.h) I've fixed - it's just adding
the trait to enable the old behavior again.