Fix infinite recursion. Simplify fallback logger.
The fallback logger now returns a valid spdlog::logger so we can
dispense with all the null checks against the underlying spdlog::logger
shared_ptr.

The fallback logger will attempt to construct a console logger. If this
fails, it constructs a logger with a null sink instead. (Neither of
these should fail, and if they do you've probably got bigger problems on
your hands.)